Move `UTxOSelection` type to `cardano-coin-selection` library.
