+++ lib/primitive/cardano-wallet-primitive.cabal
@@ -125,8 +125,6 @@ library
     Cardano.Wallet.Primitive.Types.UTxOIndex
     Cardano.Wallet.Primitive.Types.UTxOIndex.Gen
     Cardano.Wallet.Primitive.Types.UTxOIndex.Internal
-    Cardano.Wallet.Primitive.Types.UTxOSelection
-    Cardano.Wallet.Primitive.Types.UTxOSelection.Gen
     Cardano.Wallet.Unsafe
     Cardano.Wallet.Util
     Control.Monad.Random.NonRandom
@@ -186,7 +184,6 @@ test-suite test
     Cardano.Wallet.Primitive.Types.TokenQuantitySpec
     Cardano.Wallet.Primitive.Types.TxSpec
     Cardano.Wallet.Primitive.Types.UTxOIndexSpec
-    Cardano.Wallet.Primitive.Types.UTxOSelectionSpec
     Cardano.Wallet.Primitive.Types.UTxOSpec
     Data.QuantitySpec
     Spec
